requirements.txt
Flask==2.3.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
umap-learn==0.5.5
hdbscan==0.8.33
matplotlib==3.8.4
seaborn==0.13.2
tensorflow==2.16.1
